Move ref definition parsing to parser
